LPS-132250 Move PortalInstancesPortletKeys to api module
